+++ test/hover/init.lua
@@ -1699,3 +1699,18 @@ print(b.<?x?>)
 [[
 field A.x: any
 ]]
+
+TEST [[
+---@class A
+---@field x number
+---@field y number
+
+---@type A<string, number>
+local <?t?>
+]]
+[[
+local t: A<string, number> {
+    x: number,
+    y: number,
+}
+]]